Одинацатое задание (доработка, выделение настроек в отдельный класс)
